server/requirements.txt
flask
psycopg2-binary
Flask-SQLAlchemy
flask-cors
Flask-CORS
flask-login
flask_bcrypt
werkzeug
datetime
wraps
uuid
python-dateutil
paramiko
flask_jwt_extended
dotenv
grpcio-tools
grpcio-reflection
celery
redis